+++ ui/src/pages/Converter/Base64/components/ValidateStatus.tsx
@@ -1,0 +1,21 @@
+import { Button, Space } from "antd";
+import { Check, X } from "lucide-react";
+import React from "react";
+import style from "pages/Converter/Base64/Base64.module.scss";
+const ValidateStatus: React.FC<{ status: string }> = ({ status }) => {
+	return (
+		<Space className={style.base__base64Container__validator}>
+			{status.length === 0 ? null : status === "valid" ? (
+				<Button size="small" style={{ borderColor: "green" }}>
+					<Check color="green" size={16} />
+				</Button>
+			) : (
+				<Button size="small" danger>
+					<X size={16} />
+				</Button>
+			)}
+		</Space>
+	);
+};
+
+export default ValidateStatus;
