+++ ui/src/pages/Converter/Base64/index.tsx
@@ -1,17 +1,19 @@
-import { Alert, Button, Input, Space } from "antd";
+import { Button, Input, Space } from "antd";
 import { useEffect, useState } from "react";
 import { Buffer } from "buffer";
 import style from "./Base64.module.scss";
 import Clipboard from "components/RenderProps/Clipboard/Clipboard";
 import ClipboardButton from "components/General/ClipboardButton/ClipboardButton";
-import { Check, X } from "lucide-react";
+
+import ValidateStatus from "./components/ValidateStatus";
+import { isBase64Valid } from "./utils/helper";
 
 const { TextArea } = Input;
 
 const Base64: React.FC = () => {
 	const [input, setInput] = useState("");
 	const [result, setResult] = useState("");
-	const [isValidBase64, setIsValidBase64] = useState<string>("");
+	const [status, setStatus] = useState<string>("");
 
 	const onClick = (type: string, value: string) => {
 		if (type === "") return;
@@ -26,16 +28,7 @@ const Base64: React.FC = () => {
 	const IS_RESULT_EMPTY = result.length === 0;
 
 	useEffect(() => {
-		function checkValidBase64(input: string) {
-			const base64Pattern =
-				/^([A-Za-z0-9+/]{4})*([A-Za-z0-9+/]{4}|[A-Za-z0-9+/]{3}=|[A-Za-z0-9+/]{2}==)$/;
-
-			setIsValidBase64(
-				base64Pattern.test(input.trim()) ? "valid" : "invalid"
-			);
-		}
-
-		checkValidBase64(result);
+		setStatus(isBase64Valid(result));
 	}, [result]);
 
 	return (
@@ -73,18 +66,7 @@ const Base64: React.FC = () => {
 					autoSize={{ minRows: 10 }}
 				/>
 
-				<Space className={style.base__base64Container__validator}>
-					{isValidBase64.length === 0 ? null : isValidBase64 ===
-					  "valid" ? (
-						<Button size="small" style={{ borderColor: "green" }}>
-							<Check color="green" size={16} />
-						</Button>
-					) : (
-						<Button size="small" danger>
-							<X size={16} />
-						</Button>
-					)}
-				</Space>
+				<ValidateStatus status={status} />
 			</div>
 
 			<Space>
