+++ ui/src/pages/Data/ImageGeneratorFromColors/tests/imageGenerator.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from "@testing-library/react";
 import { describe, test } from "vitest";
 import user from "@testing-library/user-event";
-import ImageGeneratorFromColors from ".";
+import ImageGeneratorFromColors from "..";
 
 describe("Image Generator component", () => {
 	test("render a component without crash", () => {
