+++ ui/src/pages/Data/ImageGeneratorFromColors/utils/helper.ts
